pyproject.toml
[tool.poetry]
name = "arakna"
version = "0.1.0"
description = "Self-similar traffic research"
authors = ["Roman Voronkin <roman.voronkin@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
click = "^8.0.1"
Flask = "^2.0.1"
Jinja2 = "^3.0.1"
Werkzeug = "^2.0.1"
itsdangerous = "^2.0.1"
MarkupSafe = "^2.0.1"
numpy = "^1.21.2"
pandas = "^1.3.3"
pyarrow = "^5.0.0"
matplotlib = "^3.4.3"
scikit-learn = "^0.24.2"
joblib = "^1.0.1"
scipy = "^1.7.1"
simpy = "^4.0.1"
PyYAML = "^5.4.1"
tqdm = "^4.62.2"
xgboost = "^1.4.2"
catboost = "^0.26.1"
lightgbm = "^3.2.1"
plotly = "^5.3.1"
types-PyYAML = "^5.4.10"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.9b0"
mypy = "^0.910"
flake8 = "^3.9.2"
isort = "^5.9.3"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
# make it compatible with black
profile = "black"
